Carry the updateHosts to util.js
